Update ID type fromt Integer to Long

+++ src/main/java/ec/com/expensys/persistence/repository/ExpCountryRepository.java
@@ -3,5 +3,5 @@ package ec.com.expensys.persistence.repository;
 import ec.com.expensys.persistence.entity.ExpCountry;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ExpCountryRepository extends CrudRepository<ExpCountry, Integer> {
+public interface ExpCountryRepository extends CrudRepository<ExpCountry, Long> {
 }
